+++ lib/api/TrackEvent.ts
@@ -58,7 +58,7 @@ export class TrackApi implements ITrack {
     settings: SettingsModel,
     eventName: string,
     context: ContextModel,
-    eventProperties: any,
+    eventProperties: Record<string, dynamic>,
     hooksService: IHooksService,
   ): Promise<Record<string, boolean>> {
     if (doesEventBelongToAnyFeature(eventName, settings)) {