+++ lib/constants/Url.ts
@@ -19,4 +19,3 @@ export const HTTPS = 'https';
 export const SEED_URL = 'https://vwo.com';
 export const HTTP_PROTOCOL = `${HTTP}://`;
 export const HTTPS_PROTOCOL = `${HTTPS}://`;
-export const BASE_URL = 'dev.visualwebsiteoptimizer.com';