+++ lib/enums/UrlEnum.ts
@@ -14,10 +14,6 @@
  * limitations under the License.
  */
 export enum UrlEnum {
-  BASE_URL = 'dev.visualwebsiteoptimizer.com',
-  SETTINGS_URL = '/server-side/settings',
-  // WEBHOOK_SETTINGS_URL = '/server-side/pull',
-  // BATCH_EVENTS = '/server-side/batch-events',
   EVENTS = '/events/t',
   ATTRIBUTE_CHECK = '/check-attribute',
   GET_USER_DATA = '/get-user-details',